Replace more types

+++ Sources/Engine/Ska/Mesh.cpp
@@ -635,7 +635,7 @@ void CMesh::Write_t(CTStream *ostrFile)
           (*ostrFile)<<fFloat;
         }
         // write shader flags
-        ULONG ulFlags = msrf.msrf_ShadingParams.sp_ulFlags;
+        uint32_t ulFlags = msrf.msrf_ShadingParams.sp_ulFlags;
         (*ostrFile)<<ulFlags;
       }
     }
@@ -849,7 +849,7 @@ void CMesh::Read_t(CTStream *istrFile)
         }
         // there were no flags in shader before ver 12
         if(iFileVersion>11) {
-          ULONG ulFlags;
+          uint32_t ulFlags;
           (*istrFile)>>ulFlags;
           pshpShaderParams->sp_ulFlags = ulFlags;
         } else {
